requirements.txt
gunicorn==19.6.0	
tesserocr==2.4.0
scipy==1.3.1
pytesseract==0.3.0
pillow==6.2.1
pandas==0.25.2
opencv==3.4.1
numpy==1.17.3
Flask==1.1.1